requirements.txt
streamlit
pandas
spacy
PyPDF2
joblib
matplotlib
seaborn
wordcloud
python-docx
scikit-learn
flask 
numpy